JSON: Use extension mechanism correctly
and define all functions to be extended in the "base" module, rather
than defining new functions inside the extension and making them
accessible from the base package.
