+++ ext/SerdeJSONExt/De.jl
@@ -1,19 +1,18 @@
 module DeJson
 
-export deser_json
-
 using ..ParJson
 import Serde.to_deser
+import Serde
 
-function deser_json(::Type{T}, x; kw...) where {T}
-    return to_deser(T, parse_json(x; kw...))
+function Serde.JSON.deser_json(::Type{T}, x; kw...) where {T}
+    return to_deser(T, Serde.JSON.parse_json(x; kw...))
 end
 
-deser_json(::Type{Nothing}, _) = nothing
-deser_json(::Type{Missing}, _) = missing
+Serde.JSON.deser_json(::Type{Nothing}, _) = nothing
+Serde.JSON.deser_json(::Type{Missing}, _) = missing
 
-function deser_json(f::Function, x; kw...)
-    object = parse_json(x; kw...)
+function Serde.JSON.deser_json(f::Function, x; kw...)
+    object = Serde.JSON.parse_json(x; kw...)
     return to_deser(f(object), object)
 end
