+++ ext/SerdeJSONExt/Ser.jl
@@ -1,10 +1,9 @@
 module SerJson
 
-export to_json
-export to_pretty_json
-
 using Dates
-using Serde
+import Serde
+import Serde.JSON: isnull, ser_name, ser_type, ser_type, ser_ignore_field, ser_ignore_null,
+    ser_value
 using UUIDs
 
 const JSON_NULL = "null"
@@ -161,19 +160,6 @@ function json_value!(buf::IOBuffer, f::Function, val::AbstractSet; l::Int64, kw.
     return print(buf, indent(l - 1), "]")
 end
 
-(isnull(::Any)::Bool) = false
-(isnull(v::Missing)::Bool) = true
-(isnull(v::Nothing)::Bool) = true
-(isnull(v::Float64)::Bool) = isnan(v) || isinf(v)
-
-(ser_name(::Type{T}, k::Val{x})::Symbol) where {T,x} = Serde.ser_name(T, k)
-(ser_value(::Type{T}, k::Val{x}, v::V)) where {T,x,V} = Serde.ser_value(T, k, v)
-(ser_type(::Type{T}, v::V)) where {T,V} = Serde.ser_type(T, v)
-
-(ser_ignore_field(::Type{T}, k::Val{x})::Bool) where {T,x} = Serde.ser_ignore_field(T, k)
-(ser_ignore_field(::Type{T}, k::Val{x}, v::V)::Bool) where {T,x,V} = ser_ignore_field(T, k)
-(ser_ignore_null(::Type{T})::Bool) where {T} = false
-
 function json_value!(buf::IOBuffer, f::Function, val::T; l::Int64, kw...)::Nothing where {T}
     next = iterate(f(T))
     print(buf, "{", indent(l))
@@ -200,13 +186,13 @@ function json_value!(buf::IOBuffer, val::T; l::Int64, kw...)::Nothing where {T}
     return json_value!(buf, fieldnames, val; l = l, kw...)
 end
 
-function to_json(x...; kw...)::String
+function Serde.JSON.to_json(x...; kw...)::String
     buf = IOBuffer()
     json_value!(buf, x...; l = -1, kw...)
     return String(take!(buf))
 end
 
-function to_pretty_json(x...; kw...)::String
+function Serde.JSON.to_pretty_json(x...; kw...)::String
     buf = IOBuffer()
     json_value!(buf, x...; l = 1, kw...)
     return String(take!(buf))
